PN-16002

+++ PAS_Database/dbo/User Defined Types/POPartsToReceive.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TYPE [dbo].[POPartsToReceive] AS TABLE (
-    [PurchaseOrderId]           BIGINT NULL,
-    [PurchaseOrderPartRecordId] BIGINT NULL,
-    [QtyToReceive]              INT    NULL);
+    [PurchaseOrderId]           BIGINT          NULL,
+    [PurchaseOrderPartRecordId] BIGINT          NULL,
+    [QtyToReceive]              DECIMAL (18, 6) NULL);
